# docs/swagger.yaml
basePath: /
definitions:
  github_com_AntonMilienkov_REST-service_internal_model.Subscription:
    properties:
      created_at:
        type: string
      end_date:
        example: 12-2025
        type: string
      id:
        type: string
      price:
        example: 400
        type: integer
      service_name:
        example: Yandex Plus
        type: string
      start_date:
        example: 07-2025
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  internal_handler.errorResponse:
    properties:
      error:
        type: string
    type: object
  internal_handler.totalCostResponse:
    properties:
      total:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST-сервис агрегации данных об онлайн-подписках пользователей.
  title: Subscriptions API
  version: "1.0"
paths:
  /subscriptions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_AntonMilienkov_REST-service_internal_model.Subscription'
            type: array
      summary: Список всех подписок
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные подписки
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_AntonMilienkov_REST-service_internal_model.Subscription'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_AntonMilienkov_REST-service_internal_model.Subscription'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      summary: Создать подписку
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      summary: Удалить подписку
      tags:
      - subscriptions
    get:
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_AntonMilienkov_REST-service_internal_model.Subscription'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      summary: Получить подписку по id
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: string
      - description: Данные подписки
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_AntonMilienkov_REST-service_internal_model.Subscription'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_AntonMilienkov_REST-service_internal_model.Subscription'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      summary: Обновить подписку
      tags:
      - subscriptions
  /subscriptions/total-cost:
    get:
      parameters:
      - description: Начало периода, формат MM-YYYY
        in: query
        name: period_from
        required: true
        type: string
      - description: Конец периода, формат MM-YYYY
        in: query
        name: period_to
        required: true
        type: string
      - description: Фильтр по ID пользователя
        in: query
        name: user_id
        type: string
      - description: Фильтр по названию сервиса
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.totalCostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      summary: Суммарная стоимость подписок за период
      tags:
      - subscriptions
swagger: "2.0"